orangefs: create uapi interface
This formally separates userspace visible and kernel internal data
structures.  The sole consumer of this would be the pvfs2-client-core,
which currently simply defines all these structures itself.  Most of
these are shared with the rest of the OrangeFS userspace client and
server code.

As the userspace OrangeFS client and server progress, they are beginning
to change their "internal" definitions.  This will allow us to separate
data structures which will be sent to the kernel and data structures
which belong to userspace.

This is accomplished by moving everything relevant from fs/orangefs/*.h
to include/uapi/linux/orangefs.h.

Signed-off-by: Martin Brandenburg <martin@omnibond.com>